requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
numpy==2.2.3
pillow==11.1.0
pydexcom==0.4.1
python-dotenv==1.0.1
requests==2.32.3
RPi.GPIO==0.7.1
spidev==3.6
urllib3==2.3.0
